requirements.txt
APScheduler==3.6.3
arabic-reshaper==2.1.3
argcomplete==1.10.0
asgiref==3.3.4
beautifulsoup4==4.8.0
CacheControl==0.12.10
cachetools==5.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
countries==0.2.0
Django==3.2.2
document==1.0
docx==0.2.4
docx2pdf==0.1.7
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
ffmpeg==1.4
fire==0.4.0
firebase-admin==5.2.0
fpdf==1.7.2
future==0.18.2
geographiclib==1.50
geolocator==0.1.1
geopy==2.1.0
google-api-core==2.7.1
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.2.3
google-cloud-firestore==2.3.4
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
googletrans==3.1.0a0
grpcio==1.44.0
grpcio-status==1.44.0
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
IMAPClient==2.1.0
lxml==4.6.3
msgpack==1.0.3
numpy==1.20.2
oauth2client==4.1.3
oauth2clientziead @ git+https://github.com/zieadshabkalieh/zieadshabkalieh-oauth2clientziead@b2949e4a6275a3e9979525ec1473752c110d7f96
olefile==0.46
packaging==21.3
pdf-writer==0.1
pdf2docx==0.5.1
pdf2image==1.16.0
pdfminer.six==20181108
Pillow==8.2.0
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydriveziead @ git+https://github.com/zieadshabkalieh/zieadshabkalieh-pydriveziead@5360754fc9370e979db35ecf4f568356daa127bd
pydub==0.25.1
pyglet==1.5.18
PyMuPDF==1.18.13
pyparsing==3.0.7
pyPdf==1.13
PyPDF2==1.26.0
PySocks==1.7.1
pytesseract==0.3.7
python-bidi==0.4.2
python-docx==0.8.10
python-pptx==0.6.18
python-telegram-bot==13.4.1
pytz==2021.1
PyYAML==6.0
reportlab==3.5.67
requests==2.22.0
rfc3986==1.5.0
rsa==4.8
scipy==1.6.3
sniffio==1.2.0
somepackage==1.2.3
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
termcolor==1.1.0
textract==1.6.3
tornado==6.1
tqdm==4.60.0
uritemplate==4.1.1
xlrd==1.2.0
XlsxWriter==1.4.1
